model server/requirements.txt
fastapi
psycopg2
ultralytics
torchvision
pillow
annoy
requests
beautifulsoup4
selenium
webdriver_manager
uvicorn 
opencv-python-headless
python-multipart
boto3